[KNIFE_EC2-87] Integrate WinRM bootstrap for Windows instance
This commit adds the ability to bootstrap Windows instances via WinRM
rather than SSH.

Windows instances are detected via an EC2 API call.  When a Windows
instance is detected, the bootstrap method is choosen based on the new
bootstrap_protocol configuration option.  This option can be
overridden with the --bootstrap-protocol command line option.  The
--bootstrap-protocol option is ignored for non-Windows nodes.

The administrator password for the Windows instance is retrieved from
the EC2 API and decrypted with the specified private key.

Additional tests are still required for these features.