[KNIFE_EC2-93] Ensure ec2 hint isn't set simply by loading the plugin.
Currently, the Chef::Config[:knife][:hints]["ec2"] option will be set
simply by loading the plugin, causing the hint file to be erroneously
created on non-ec2 nodes.  This commit ensure we only set the option
in the #run method or in the Proc that is called when the --hint
option is used.