Remove instance colors
We have hardcoded instance colors that show up in knife ec2 server list.
The colors are hard coded to incredibly old instance types.  We *could*
update this for the current instance types, but it would be outdated in
another quarter.  It's probably better to just remove this functionality
since it's confusing as-is.