Loosen deps on fog-aws and rake
The new fog-aws (1.1) introduces support for additional regions and
instance sizes. We should make sure we pull that in so we can support
those as well. It doesn't seem like they actually broke anything when
they went 1.0.

While I was here I added testing on Ruby 2.4 and removed branch testing
to avoid double testing PRs. We do that everywhere chef/*

Signed-off-by: Tim Smith <tsmith@chef.io>