Verify windows password is part of the provided options
WinRM config data is passed in the user data to the AWS api.
Part of this includes an optional winrm user creation.
The user can only be created if a password is passed to the gem.
it cannot wait for the password from AWS, as this is only available
after the user data is paased and the server created

Signed-off-by: Topher Cullen <topher.cullen@jamberry.com>