Fallback to the default credentials file if it exists
AWS keys on the CLI and in knife.rb will continue to take precedent, but if you specify nothing on the CLI now we'll go looking for a credentials file and if we find it we'll use those credentials. In 2016 people should be using the credentials file unless they have a need to override those values so this makes that easy.

Signed-off-by: Tim Smith <tsmith@chef.io>