Fix a few failure conditions that could occur
We can't double validate the config since it loads the values when we do that and then we fail the config file + keys scenario later on.

side note: why does validation load config values into the config hash. That seems like a bad idea

Make sure to load the correct config in the credentials loading method

Signed-off-by: Tim Smith <tsmith@chef.io>