Add connection debug information
I was having a really hard time getting this plugin to work on my
system. I thought I could run the plugin in debug, but that didn't
provide any connection information. I added debug statements to various
parts of the connection setup and I improved the error when the aws
profile is missing from your credentials file, which is what my error
actually was. This will result in secret keys being written to the
terminal, but there isn't really a way around that when someone runs a
cloud plugin in debug.

Signed-off-by: Tim Smith <tsmith@chef.io>